+++ client-common/src/test/java/org/realityforge/replicant/client/transport/TestDataLoadService.java
@@ -95,7 +95,7 @@ final class TestDataLoadService
 
   @Nonnull
   @Override
-  protected Class<TestGraph> getGraphType()
+  public Class<TestGraph> getGraphType()
   {
     return TestGraph.class;
   }